Implementación del test BeAbleToAddAllEarningsToExistingDictionaryOfEarningsPerProviderGivenFromAndToDates() en la clase PaymentShould. Estado: No compila.

+++ TodoPagos/TodoPagos.Web.Api.Tests/EarningQueriesControllerShould.cs
@@ -37,7 +37,7 @@ namespace TodoPagos.Web.Api.Tests
             DateTime from = DateTime.ParseExact("Mon, 15 Sep 2008 09:30:41 GMT", 
                 "ddd, dd MMM yyyy HH':'mm':'ss 'GMT'", CultureInfo.InvariantCulture);
             DateTime to = DateTime.Today;
-            IDictionary<Provider, int> result = new Dictionary<Provider, int>();
+            IDictionary<Provider, double> result = new Dictionary<Provider, double>();
             result.Add(new Provider("Antel", 10, new List<IField>()), 100);
             result.Add(new Provider("Tienda Inglesa", 7, new List<IField>()), 200);
             var mockEarningQueriesService = new Mock<IEarningQueriesService>();
@@ -45,8 +45,8 @@ namespace TodoPagos.Web.Api.Tests
             EarningQueriesController controller = new EarningQueriesController(mockEarningQueriesService.Object);
 
             IHttpActionResult actionResult = controller.GetEarningsPerProvider(from, to);
-            OkNegotiatedContentResult<IDictionary<Provider, int>> contentResult = 
-                (OkNegotiatedContentResult<IDictionary<Provider, int>>)actionResult;
+            OkNegotiatedContentResult<IDictionary<Provider, double>> contentResult = 
+                (OkNegotiatedContentResult<IDictionary<Provider, double>>)actionResult;
 
             Assert.AreSame(contentResult.Content, result);
         }
@@ -57,7 +57,7 @@ namespace TodoPagos.Web.Api.Tests
             DateTime from = DateTime.ParseExact("Wed, 29 Aug 1962 00:00:00 GMT",
                  "ddd, dd MMM yyyy HH':'mm':'ss 'GMT'", CultureInfo.InvariantCulture);
             DateTime to = DateTime.Today;
-            IDictionary<Provider, int> result = new Dictionary<Provider, int>();
+            IDictionary<Provider, double> result = new Dictionary<Provider, double>();
             result.Add(new Provider("Antel", 10, new List<IField>()), 100);
             result.Add(new Provider("Tienda Inglesa", 7, new List<IField>()), 200);
             var mockEarningQueriesService = new Mock<IEarningQueriesService>();
@@ -65,8 +65,8 @@ namespace TodoPagos.Web.Api.Tests
             EarningQueriesController controller = new EarningQueriesController(mockEarningQueriesService.Object);
 
             IHttpActionResult actionResult = controller.GetEarningsPerProvider();
-            OkNegotiatedContentResult<IDictionary<Provider, int>> contentResult =
-                (OkNegotiatedContentResult<IDictionary<Provider, int>>)actionResult;
+            OkNegotiatedContentResult<IDictionary<Provider, double>> contentResult =
+                (OkNegotiatedContentResult<IDictionary<Provider, double>>)actionResult;
 
             Assert.AreSame(contentResult.Content, result);
         }
